* OpenGLShader code move and gtk warning fixed
* char* => std::string for shader names and cleanup
* fixed a gtk null ptr warning in the texture browser

+++ src/tools/radiant/include/iglrender.h
@@ -47,7 +47,7 @@ class OpenGLState {
 			eSortLast = 4096
 		};
 
-		int m_polygonOffset;
+		float m_polygonOffset;
 		unsigned int m_state;
 		std::size_t m_sort;
 		GLint m_texture;
